pyproject.toml
[project]
name = "git-analyzer"
version = "0.1.0"
description = "A TUI git repository analyzer"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "textual>=0.47.1",
    "gitpython>=3.1.40",
    "click>=8.1.7",
    "textual-plotext>=0.2.1",
]

[project.scripts]
git-analyzer = "git_analyzer.main:main"
git-analyzer-dev = "textual run --dev git_analyzer.main:main"
